fix lint issue
